rename to generic.dispatch
