+++ docs/source/pyro.generic.rst
@@ -4,7 +4,7 @@ Generic Interface
 The ``pyro.generic`` module provides an interface to dynamically dispatch Pyro code
 to custom backends.
 
-.. automodule:: pyro.generic.generic
+.. automodule:: pyro.generic.dispatch
     :members:
     :undoc-members:
     :show-inheritance:
