+++ pyro/generic/__init__.py
@@ -1,4 +1,4 @@
-from pyro.generic.generic import distributions, handlers, infer, ops, optim, pyro, pyro_backend
+from pyro.generic.dispatch import distributions, handlers, infer, ops, optim, pyro, pyro_backend
 
 __all__ = [
     'distributions',
